[UDAF] Vectorization support remote udaf #10683
